+++ src/app/pages/PerpetualPage/hooks/usePerpetual_OpenPositions.ts
@@ -11,7 +11,7 @@ import {
 } from '../types';
 import { Event, useGetTraderEvents } from './graphql/useGetTraderEvents';
 import { useMemo } from 'react';
-import { ABK64x64ToFloat, ABK64x64ToWei } from '../utils';
+import { ABK64x64ToFloat, ABK64x64ToWei } from '../utils/contractUtils';
 import { BigNumber } from 'ethers';
 import { Nullable } from '../../../../types';
 import { usePerpetual_queryAmmState } from './usePerpetual_queryAmmState';
@@ -20,7 +20,8 @@ import {
   getMarkPrice,
   getTraderPnL,
   getBase2QuoteFX,
-} from '../temporaryUtils';
+  getBase2CollateralFX,
+} from '../utils/perpUtils';
 import { usePerpetual_queryTraderState } from './usePerpetual_queryTraderState';
 import { numberFromWei } from '../../../../utils/blockchain/math-helpers';
