+++ src/app/pages/PerpetualPage/hooks/usePerpetual_depositMarginToken.ts
@@ -5,7 +5,11 @@ import { getContract } from 'utils/blockchain/contract-helpers';
 import { toWei } from 'web3-utils';
 import { gasLimit } from 'utils/classifiers';
 import { TxType } from 'store/global/transactions-store/types';
-import { PERPETUAL_ID, floatToABK64x64, checkAndApprove } from '../utils';
+import {
+  PERPETUAL_ID,
+  floatToABK64x64,
+  checkAndApprove,
+} from '../utils/contractUtils';
 
 export const usePerpetual_depositMarginToken = () => {
   const account = useAccount();
