+++ src/app/pages/PerpetualPage/hooks/usePerpetual_queryPerpParameters.ts
@@ -2,8 +2,8 @@ import { bridgeNetwork } from 'app/pages/BridgeDepositPage/utils/bridge-network'
 import { useEffect, useState } from 'react';
 import { Chain } from 'types';
 import { getContract } from 'utils/blockchain/contract-helpers';
-import { PerpParameters } from '../temporaryUtils';
-import { ABK64x64ToFloat, PERPETUAL_ID } from '../utils';
+import { PerpParameters } from '../utils/perpUtils';
+import { ABK64x64ToFloat, PERPETUAL_ID } from '../utils/contractUtils';
 import perpetualManagerAbi from 'utils/blockchain/abi/PerpetualManager.json';
 
 const initialPerpParameters: PerpParameters = {
