+++ src/app/pages/PerpetualPage/utils/perpUtils.ts
